requirements.txt
numpy==1.20.3
object_detection==0.1
opencv_python_headless==4.5.5.62
python_firebase==1.2.1
tensorflow==2.7.0
